feat: add MVP route feature flag and update layout components
