+++ web/.env.example
@@ -4,6 +4,7 @@ BETTER_AUTH_URL=http://localhost:3000
 NEXT_PUBLIC_APP_URL=http://localhost:3000
 NEXT_PUBLIC_DISABLE_ACCOUNT_ROUTE=false
 NEXT_PUBLIC_DISABLE_AGENTS_ROUTE=false
+NEXT_PUBLIC_DISABLE_MVP_ROUTE=false
 TRUSTED_ORIGINS=http://localhost:3000
 IRONHUB_KEY_ENCRYPTION_SECRET=replace-with-32-byte-random-secret
